+++ src/loading.jl
@@ -1,38 +1,27 @@
 import ..ClimaComms
 
-export import_required_backends
+"""
+    mpi_is_required()
 
-function mpi_is_required()
-    return context_type() == :MPICommsContext
-end
+Returns a Bool indicating if MPI should be loaded, based on the
+`ENV["CLIMACOMMS_CONTEXT"]`. See [`ClimaComms.context`](@ref) for
+more information.
 
-function mpi_ext_is_not_loaded()
-    return isnothing(Base.get_extension(ClimaComms, :ClimaCommsMPIExt))
-end
-
-function cuda_is_required()
-    return device_type() == :CUDADevice
-end
-
-function cuda_ext_is_not_loaded()
-    return isnothing(Base.get_extension(ClimaComms, :ClimaCommsCUDAExt))
-end
+```julia
+mpi_is_required() && using MPI
+```
+"""
+mpi_is_required() = context_type() == :MPICommsContext
 
 """
-    ClimaComms.@import_required_backends
+    cuda_is_required()
 
-If the desired context is MPI (as determined by `ClimaComms.context()`), try loading MPI.jl.
-If the desired device is CUDA (as determined by `ClimaComms.device()`), try loading CUDA.jl.
+Returns a Bool indicating if CUDA should be loaded, based on the
+`ENV["CLIMACOMMS_DEVICE"]`. See [`ClimaComms.device`](@ref) for
+more information.
+
+```julia
+cuda_is_required() && using CUDA
+```
 """
-macro import_required_backends()
-    return quote
-        @static if $mpi_is_required() && $mpi_ext_is_not_loaded()
-            import MPI
-            @info "Loaded MPI.jl"
-        end
-        @static if $cuda_is_required() && $cuda_ext_is_not_loaded()
-            import CUDA
-            @info "Loaded CUDA.jl"
-        end
-    end
-end
+cuda_is_required() = device_type() == :CUDADevice